Loading OBJs
Now that I've sort of completed the loader it was time to test it. Sadly it failed. I imagine it has something to do with the fact that the memory arena size isn't big enough. But as I'll restate from my comment where the fault occurs. 'I imagine it has something to do with the fact of the size of the memory arena, but I was too dumb or lazy to make a better error checking system'. It has now come back to bite me in the butt.